Day-05/project/pyproject.toml
[project]
name = "agent-ops-day3"
version = "0.1.0"
description = "AI Agent Ops Closed-Loop System (Dataset → Evaluation → Remediation)"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "deepagents>=0.2.7",
  "langchain>=1.2.3",
  "langchain-openai>=1.1.7",
  "langgraph>=1.0.6",
  "langfuse>=3.11.2",
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.2,<3",
  "python-dotenv>=1.0.0",
  "structlog>=25,<26",
]

# PEP 735: 개발/테스트 의존성 분리
[dependency-groups]
dev = [
  "deepeval>=3.8,<4",
  "mypy>=1.19.1",
  "pandas>=2.2,<3",
  "pytest>=9,<10",
  "pytest-asyncio>=1.0,<2",
  "ruff>=0.15.2",
]

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "src"
module-root = ""

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
